+++ src/Pim/Bundle/FlexibleEntityBundle/Model/AbstractFlexibleValue.php
@@ -104,9 +104,15 @@ abstract class AbstractFlexibleValue implements FlexibleValueInterface, Localiza
      * @param AbstractAttribute $attribute
      *
      * @return AbstractFlexibleValue
+     * @throws LogicException
      */
     public function setAttribute(AbstractAttribute $attribute = null)
     {
+        if (is_object($this->attribute) && ($attribute != $this->attribute)) {
+            throw new \LogicException(
+                sprintf('An attribute (%s) has already been set for this value', $this->attribute->getCode())
+            );
+        }
         $this->attribute = $attribute;
 
         return $this;